Adding an outer loop and a wait to reduce requests
At this point I don't think the check for an invalid token is even worth it, since if it's not "valid", it'll retrieve a new token anyway...

Anyhow, don't have time to test that, so just going with what I know works right now. Also got rate limited for once so added a small timer to reduce the amount of hits.